added casts to avoid warnings

+++ src/PDFViewBookmarks.cpp
@@ -55,7 +55,7 @@ public:
 		}
 		if (dest)
 		{
-			unsigned long pageIndex = FPDFDest_GetDestPageIndex(doc, dest);
+			int pageIndex = (int) FPDFDest_GetDestPageIndex(doc, dest);
 			pdfView->GoToPage(pageIndex);
 		}
 	}
